+++ app/src/pages/EvalLogSearchPage/EvalLogSearchDetail.tsx
@@ -1,7 +1,7 @@
 import { EvalLogEdge } from '@/__generated__/graphql';
 import { ApolloError } from '@apollo/client';
 import { Center, Skeleton, Text, VStack } from '@components/common';
-import { ApolloBadRequest } from '@components/elements/DashboardContentView';
+import { ApolloBadRequest } from '@components/elements/DashboardContentView/Error';
 import { Footer } from '@components/elements/Footer/Footer';
 import { useTheme } from '@emotion/react';
 import React from 'react';
