+++ app/src/pages/HomePage/contents/Coalition/ScoreRecordsPerCoalition.tsx
@@ -5,7 +5,7 @@ import { LineChart } from '@components/elements/Chart';
 import {
   ApolloBadRequest,
   ApolloNotFound,
-} from '@components/elements/DashboardContentView';
+} from '@components/elements/DashboardContentView/Error';
 import { DashboardContent } from '@components/templates/DashboardContent';
 import { numberWithUnitFormatter } from '@utils/formatters';
 import { isDefined } from '@utils/isDefined';
