+++ app/src/pages/HomePage/contents/User/AverageDurationPerCircle.tsx
@@ -5,7 +5,7 @@ import { HorizontalBarChart } from '@components/elements/Chart';
 import {
   ApolloBadRequest,
   ApolloNotFound,
-} from '@components/elements/DashboardContentView';
+} from '@components/elements/DashboardContentView/Error';
 import { DashboardContent } from '@components/templates/DashboardContent';
 import { numberWithUnitFormatter } from '@utils/formatters';
