+++ app/src/pages/HomePage/contents/User/WeeklyEvalCount.tsx
@@ -5,7 +5,7 @@ import { Loader } from '@components/common';
 import {
   ApolloBadRequest,
   ApolloNotFound,
-} from '@components/elements/DashboardContentView';
+} from '@components/elements/DashboardContentView/Error';
 import { NumberCompare } from '@components/elements/DashboardContentView/Text';
 import { DashboardContent } from '@components/templates/DashboardContent';
 import dayjs from 'dayjs';
