+++ app/src/pages/ProfileGeneralTab/contents/BeginAt.tsx
@@ -4,7 +4,7 @@ import { H3Text, HStack, Loader, Text } from '@components/common';
 import {
   ApolloBadRequest,
   ApolloNotFound,
-} from '@components/elements/DashboardContentView';
+} from '@components/elements/DashboardContentView/Error';
 import { DashboardContent } from '@components/templates/DashboardContent';
 import dayjs from 'dayjs';
 import { useParams } from 'react-router-dom';
