+++ app/src/pages/ProfileGeneralTab/contents/PreferredTime.tsx
@@ -5,7 +5,7 @@ import { H3Text, HStack, Loader, Text, VStack } from '@components/common';
 import {
   ApolloBadRequest,
   ApolloNotFound,
-} from '@components/elements/DashboardContentView';
+} from '@components/elements/DashboardContentView/Error';
 import { ProgressionBar } from '@components/elements/ProgressionBar';
 import { DashboardContent } from '@components/templates/DashboardContent';
 import dayjs from 'dayjs';
